Many fixes
Accept now working

+++ src/lreader.hpp
@@ -5,7 +5,7 @@
 
 const int ReaderEof = -1;
 
-const int ReaderBufferSize = 66000;
+const unsigned long ReaderBufferSize = 66000;
 
 class LReader {
 protected:
